refactor: move amqp logic to engineConnector
