+++ packages/engine-server/lib/index.js
@@ -9,5 +9,4 @@ const server = require('http').createServer()
 const io = require('socket.io')(server)
 server.listen(port)
 
-require('./amqp')({ amqp, io, bookingsCache, vehiclesCache, planCache })
 require('./routes')({ io, bookingsCache, vehiclesCache, planCache })
